Created HP Service Layer to improve readbility
